Keybind reset & jumping bug fix
Both bugs were actually the same thing!

The mod's config was only being loaded after `PostModsInit()`, and it wasn't being filled in with saved keybinds at all unless the Jolly Co-op menu was opened, since that's usually done *during* `PostModsInit()`.

The fix was to make `JollyRebindConfig` bind everything in its constructor which is called in `OnModsInit()` (which it should have been doing anyway really), and if it needs to be changed in `PostModsInit()`, recreating it and calling `Reload()` in order to manually fill it with any saved keybinds.